Add some excluded files removed from master
Some old, largely obsolete files had been removed from master for some reason,
so I did not reinstate them, but it seems that the compilation fails without them,
so for compatability reasons I have added them again.

At some point in the future I will remove them and fix the other files that depend
on them, but I do not want to do that now, since this change is to synchonrise the
code between 21.3 and master, so these additional changes will be made to 21.3
and then swept in once these changes have gone in and the code between the releases
is convergent.