Changes to disable TrigIDJpsi monitoring
We want to leave the code in the release for the time being, but it is
now mostly obsolete and rather out of date so we do not want it being built.

These changes hopefully remove references to it wherever it is called from
the overall monitoring, and the jet monitoring etc.
a