AthAllocators: Try to resolve lock hierarchy violation.
Re: ATLASSIM-4003.
Not sure this isn't a false positive, but ensure that we release
the AreaBase lock before waiting on an allocator lock, even in the
makeAllocator path.  (May be a false positive since no one else can be
contending for the lock in that case, as the allocator has just been created.)