Use Default List for the ART Grid Inputs in MuonRecRTT
The ART input container, user.artprod.input.container, has been updated with the default
list and the number of files used from each dataset. This MR is to use the datasets in
the default list in MuonRecRTT test scripts, that is, to add "user.artprod." prefix to
the dataset names in art-input lines.