add a set of new ISelectionAccessor implementations
The goal is to make it a lot easier to add simple transformations on
input objects, removing the need to add such logic to individual CP
algorithms performing their preselection.  Indeed some of the existing
CP algorithms that need to read such decorations already have logic in
the algorithm itself for exactly this kind of functionality.

This also avoids the issue with existing algorithms that if you have
one property with a list of selection decorations and another with a
list of invert flags, that both of those lists have to be matched up.