Turn off the addSectors flag for non-NSW reconstruction
Having this flag set to true results in the reconstruction of a large number of additional segments.
Since the motivation was to improve the reconstruction of NSW segments, and the quality of these segments has not been studied, this change turns off the use of the flag for non-NSW reco.