Improve the style of GX2F and remove dead code
This commit is an attempt to improve the coding style (purely in terms
of syntax) of the GlobalChi2Fitter to make it more readable and
maintainable. In addition, we remove a lot of dead code. Some of that is
truth code that no longer works (or even compiles) and some is code that
was commented out. We also modify the way the GlobalChi2Fitter
initializer looks to make it more inline with idiomatic C++.