add preselection option to all (relevant) algorithms
The idea here is that this allows to run CP algorithm sequences
without view containers.  So far I only touched the algorithms
themselves, but did not change the sequence configuration.  Also,
I did not update the algorithms that pass a whole container into
the tool.  That would be doable, but would require some more
coding effort.