Add new CscOverlayDigitToRDO Algorithm to convert signal CSC Digits to CSC RDOs
Moved signal CSC Digit to CSC RDO conversion from `CscOverlay` to a separate
`CscOverlayDigitToRDO` Algorithm. Currently this is just a copy-paste of the old
code, but it would be useful to know the circumstances which determine whether
`numSamples` is two or four. Perhaps this can be configured at the python level.
This would remove the need to check the background `CscRawDataContainer` to
determine how the signal CSC Digits should be processed.

Standardized `CscOverlay` input/ouput property naming at the same time.