Fix Combined MuonRecRTT tests
Jet reconstruction is turned off by the MuonCombinedRecOnlySetup, so the BeamBackgroundFiller needs to be disabled.
However, it can't be done in the MuonCombinedRecOnlySetup itself, since that's a preExec, and I couldn't quickly find a nice way to get the BeamBackgroundFiller to check if jet rec was turned on.
So I just turned it off in a postExec here.