Add enums for btag EDM schema
This makes the names of variables that FlavorTagDiscriminants reads and writes
configurable. The current names aren't very widespread but are used in a few
derivations, so I can't just change them. That said, they were lousy and
confusing names, and we can do better.

At some point we'll remove all the old names completely, but I'd like to
give a grace period.