Fix track selection for RNN
We edited the track selection for the LSTNN inputs to match
the Loose selection from the tracking group.
The only thing that we are not sure about is the # Si hits cut,
because the InDetTrackTool seems to be using the dead pixel
and SCT hits, but right now, we're not using them for the
nSiHits > 7 cut.