Make ByteStream_StorageType available via ByteStreamAddress::storageType()
ByteStream_StorageType is an external variable referenced by all
bytestream converters. Make this available via a static constant method in
ByteStreamAddress to be used by all clients. This is more explicit and
will also allow tools like the thread checker to correctly deduce that
this is a constant value.

Once all clients are migrated remove the external variable.