Extend the functionality of the PU residual calibs
New features:

- It is now possible to apply the pileup residual calibrations without
  applying the jet area-based calibration. This is needed for HLT jets
where it is not possible to re-apply the jet area-based calibrations
since the online rho is not stored. With this, we can pick-up the jets
corrected by the jet area-based calibration and appy on top the residual
calibrations. This was implemented by Florencia Daneri.

- It is possible to use nJet (number of jets above a given threshold) instead of NPV in the pile-up residual calibrations. This is needed for the TLA analysis. This was implemented by Will Kalderon.

- Sequential residual: It is now possible to apply first the mu-based
  calibration and then NPV- or nJet-based calibration. It is also
possible to apply only the mu-, NPV- or nJet-based calibrations, which
is helpful for derivingthe sequential residual calibrations.