Disabled some Eigen dictionaries for newever versions of Eigen.
The Eigen 3.3.7 version coming with LCG_95 does not come with some of
the previous underlying classes. For now I just disabled building the
dictionaries for the types that are no longer there. But we may need
to also add dictionaries for some other types as well. We'll see at
runtime.