Made the package Boost-1.67++ compatible.
With Boost 1.67.0 the library previously called boost_python
was renamed to boost_python27. This is to be able to have multiple
versions of the library side by side.

To account for this in the C++ code, the TrigConfDBConnection package
is now told to look for either the "python" or "python27" component
of Boost.