Fix for ATLASRECTS-4869
Due to a mistake on my part, the MuonPatternCombinationCollection was no longer being written to SG, leading to a leak: that is now fixed.
The fact that this was only noticed by the leak suggests that there is no longer any need for the MuonPatternCombinationCollection to be in SG at all.
If this is confirmed, it will be removed in a subsequent MR.