AthenaKernel: Fix deletion ordering problem with RecyclableDataObject.
During finalization, it is possible that a component with a
RecyclableDataQueue gets deleted while there are still associated
data objects in the event store.  This can then lead to a use-after-free
error when the store is finalized.  Be robust against this by allocating
the underlying queue separately from RecyclableDataQueue and managing
it using shared_ptr.

Should fix ATLASSIM-4044.