TrkParametersBase: Fix ubsan warnings due to cling alignment.
This package contains the declaration:

   static const Amg::Vector2D INVALID_POSITION(INVALID,INVALID);

It appears, however, that when cling generates code for this
declaration, it does not respect the proper alignment --- we were
seeing this object 8-byte aligned, rather than 16-byte as eigen
requires.  This was causing ubsan warnings as well as assertion
failures from eigen.

This variable is, however, only referenced frmo two places in the same file,
and in those places we can just as well construct is as needed.
So get rid of this static.