Fix auto-configuration for isOnline mode
The fully fledge AutoConfiguration is not usable with `isOnline=True`
outside of P1 because it actually tries to access information from IS
servers. Revert back to using `AthFile` directly as it was done in
runHLT_standalone.py

The default is still `isOnline=False` but added a command line option to
switch the mode for ease of debugging problems we are currently seeing.
Once those are fixed we should set the default to `isOnline=True`.