Sweep athenaprivate1/!15663 and athenaprivate1/!15680 from 21.3 into master.
This commit tidies the syntax of the `FastChain_tf.py` transform skeleton which is a merge
of the simulation and digitization configuration and therefore does a lot of things twice
(or more).

The eventual goal of this is to be able to refactor the skeleton, such that we set all the
required flags, then create the Algorithms, Tools and Services afterwards.