Update FastChainPileup ART jobs to match configurations from 21.3
Currently the `FastChainPileup` ART jobs are not run for the ~master branch,
but prior to activating them, then we should get the syntax properly up-to-date.

In order to have reproducible output on all Intel architectures, the
`FastChainPileup` jobs should be run without the Intel Math library.
The `art.py compare` script does not require the `--imf False` option,
so removed this.
Cleaned up whitespace and made the `art-result` strings consistent between jobs.

Fixed `postExec` statements to ensure that the configuration of `FastChain_tf.py`
ART jobs is written out to `config.txt` which should then be saved by ART.