Clean up PhysicsValidationUserAction
`PhysicsValidationUserAction` had handles to a number of tools and services
which weren't used. Dropped `ServiceHandle<ISF::IParticleBroker>`,
`ToolHandle<ISF::IParticleHelper>` and `ServiceHandle<G4UA::IUserActionSvc>`.

Dropping the dependency on the `ParticleBroker` instance removed the need for
separate configurations of the `PhysicsValidationUserActionTool` for different
simulator types. This then simplied the configuration of the `UserActionSvc`
for ISF jobs meaning that a separate `G4UA::ISFQuasiStableUserActionSvc`
was no longer necessary.