First step of the reorganization/cleanup of the LArHV package
The goal of this reorganization is to drop the dependency of the LArHV
package on custom reference counting mechanism implemented by GeoModelKernel/RCBase.
The latter mechanism is thread-unsafe and it is not really needed for LArHV.

As a first step of this reorganization, only EMBHV* classes have been modified.
The EMBHV*ConstLink typedefs have been dropped and the package clients have been
migrated to the new interface.