Fix a recursion bug in GX2F::fit(MeasSet)
This part of the code attempts to correct some points along a set of
measurements if a Micromegas detector is involved but it also introduces
a stack corruption bug that occurs in some jobs due to the function
infinitely (non-tail) recursing.

The original author attempted to provide a base case for this recursive
correction strategy but relied on a variable in the cache object which
is not passed through in the recursion. This renders the entire variable
moot and we have therefore removed it.