Remove AdjustableT0Tool (Fix for ATLASRECTS-4825)
The AdjustableT0Tool is used by two tools to cache segment t0 values in order to pass them to the MdtDriftCircleOnTrackCreator: this is obviously not MT-friendly.
The first tool that does this is the DCMathSegmentMaker: since in this case there is only one such t0, it is trivial to pass it to the MdtDriftCircleOnTrackCreator.
The other tool, the MooTrackBuilder, potentially caches more values: however, the method that does this is not used anywhere, so I removed it.
The AdjustableT0Tool can then be removed from the configuration of the MdtDriftCircleOnTrackCreator, which simply uses the t0 passed from the DCMathSegmentMaker.
(Or, in the case of stau reconstruction, the t0 it retrieves from the separate stau t0 tool.)
For now, I'm not deleting the AdjustableT0Tool entirely, as it might come in handy if rewritten to be MT-friendly, but at the moment there is no need for it.