Update to RHadron Mass Spectrum for Spin Issue
We identified an issue in the treatment of spins in the R-mesons, which
requires one more fix to the masses. Along the way, we also fixed a
rounding issue in the calculation's output which leads to a few other 1
MeV differences here and there. These changes are also in svn, and
should go ~everywhere in git.