Porting RNN tau ID to master
Hello,

This MR ports the RNN tau ID code to master. The reference code comes from 21.1 (now 21.3), which has the trigger functionalities used in 2018 tau triggers (0-prong network that is not used for offline taus).
The tuning files are the final ones, while 21.0 currently has the preliminary tunes which are then overwritten via an AODFix while making derivations.
For master, it makes sense to use the final tuning directly.

This MR had to touch some JetTagging packages, because they had a local copy of a few source/header files of the lwtnn package (version 2.0), while version 2.8 is now is master, which created some conflict.
Now, JetTagging points to the AtlasExternals library.

This MR will probably violate the frozen T0 policy because we now schedule the RNN tau ID in the tau reconstruction, which adds a couple of scores as well as some working points (the tau EDM has been updated accordingly).

Cheers,
Bertrand