Updates to R-Hadron simulation for stable particles
This is a couple of twweaks to double protect us when running stable
R-Hadron simulation. The idea is that people should be able to not set a
lifetime or set a lifetime to -1 or 0 and assume that the simulation
will guess they want stable particles. Explicitly setting doDecays=0
will also work. Also added a catch in the simulation to make sure that
the settings read in are consistent (e.g. if someone sets a lifetime,
they almost certainly want decays, and if they set a negative lifetime
they almost certainly don't).

Marking as a WIP until we can get some tests in.