Stopped xAOD::TEvent from making the first event of a file available automatically.
Until now after calling xAOD::TEvent::readFrom(...), the first event (or actually
whatever the internal m_entry variable was set to...) was automatically made
available to the user from the file. This update changes this. So that after
calling readFrom(...), only the metadata of the input file is available.
To get access to the event data, xAOD::TEvent::getEntry(...) must be called
with a valid entry number first.