Fixes for MuonTrackPerformanceAlg
Previously, the MuonTrackPerformanceAlg relied on TrackCollections holding all the tracks of the type it was interested in.
However, such TrackCollections still exist only for MS tracks.  Therefore, the code is modified to obtain the tracks it needs directly from the xAOD::MuonCollection.
New properties are added to determine the type of track (following the xAOD::Muon::TrackParticleType enum) and whether or not Staus are being processed.
For all track types except MS tracks, the track type and stau flag is used to determine both the type of muons to be used and which tracks to retrieve from them.
For compatibility with the MS track implementation and the MuonTrackTruthTool, the Trk::Track objects are copied into a unique_ptr to a TrackCollection which is then used for the remainder of the processing.
The configuration is modified accordingly, with the TrackInputLocation being blank for all cases except that of the MuonSpectrometerTracks.
Also, a bug in the use of the m_doNSW flag in initialize() was fixed: the relevant if statement should have been a check to see if the flag was false.

Additionally, a bug was fixed in the MuonDetailedTrackTruthMaker.
As "MSOnlyExtrapolatedMuonTracks" contained "ExtrapolatedMuon", the code was never making it to the check for "MSOnly": moving this to come first fixes the issue.