Remove usage of thread-unsafe random number serivce from CscDigitToCscRDOTool
This commit removes the usage of `IAtRndmGenSvc` from `CscDigitToCscRDOTool`
and replaces it with the thread-safe `IAthRNGSvc`.

These changes are a necessary step towards reproducible digitization output
between single-threaded and multi-threaded jobs. The digitization and overlay
output will change due to this commit as the random number seeding will be
slightly different.