Fix for ATLASRECTS-4909
ATLASRECTS-4909 reports a crash in the IsolationBuilder due to the ParticleCaloExtensionTool being unable to extrapolate an SA muon back through the calorimeter.
In principle, all SA muons should be extrapolated back through the calorimeter already, so this should never happen.
However, investigating this particular case showed that things are not as straightforward as I assumed.
The extrapolation is first done using a guess at the track parameters: after the fit, it is done again, the calo TSOS's updated, and the fit re-done.
In this case, though, the track fit failed after the calo TSOS's were updated.  Not only that, the final extrapolated fit had a worse chi2 than the initial MS track.
The logical fix, then, is to reject extrapolated tracks in such cases.  The corresponding MS tracks can still be used in combined fits, but not turned into SA muons.
This should affect only a handful of muons (0 in a sample of 1000 Z->mumu events, in fact).