Remove usage of thread-unsafe random number serivce from TRT_PAI_Process
This commit removes the usage of `IAtRndmGenSvc` from `TRT_PAI_Process`
and instead passes a pointer to a `CLHEP::HepRandomEngine` as a method
arguement when required.

`TRTDigitizationTool` and `TRTProcessingOfStraw` modified to create
a `CLHEP::HepRandomEngine` and pass it along to `TRT_PAI_Process`.

These changes are a necessary step towards reproducible digitization
output between single-threaded and multi-threaded jobs. The digitization
and overlay output will change due to this commit as the random number
seeding will be slightly different.