Add support for complex jet chains. Temporarily limit menu items for simple chains.
While the hypo tools are set up by a call to  TrigHLTJetHypoConfig.trigJetHypoToolFromName()
only a few chains are supported. The situation will change when
TrigHLTJetHypoConfig.trigJetHypoToolFromDict() is called

Menus have been limited to HLT_j85 and HLT_j45 for this reason:
	Trigger/TrigValidation/TrigUpgradeTest/share/jet.menu.py
	Trigger/TrigValidation/TrigUpgradeTest/share/jetMenu.ref
	Trigger/TrigValidation/TrigUpgradeTest/share/simpleJetJob.py

Failing monitoriung code for Caalorimeter (???) commented out
	Trigger/TrigValidation/TrigUpgradeTest/python/jetDefs.py

The remaining changes are C++ and python configuration files for complex chains.