LArRecConditions: Clean up LArFebConfig.
Remove mutable cache in LArFebConfig --- it can't really be making
any significant difference.  But add a method to retrieve both upper
and lower thresholds in one go.  Thus can also remove the locking.
Convert the map to an unordered_map.  Add a method to fill the map,
allowing the removal of the friend declaration.