WIP: Last import of 21.3 tauRecTools functionalities into master
Hello,

This is the last MR which migrates the trigger functionalities from 21.3 to master.
It mostly adds MVA energy scale calibration for HLT taus and allows for 0-prong taus in the working point decorator.
It will likely be considered as violating the frozen T0 policy because two decorations are added to taus for online MVA TES (`LeadClusterFrac` and `UpsilonCluster`).
On purpose, I did not migrate the TauTrackClassifier code, because these functionalities were never used in 2018 triggers, and will likely not be used for Run3 if we manage to deploy RNN track classification.

Cheers,
Bertrand