TrigTauRec: Avoid ERROR messages from undefined data handle.
A handle was not being initialized, leading to ERROR messages in debug builds.
The code did explicitly handle the case of the object not being present,
so that's apparently expected.  Update the way the handle is used
to be consistent with this.

Fixes some failures seen in TrigUpgradeTests in the debug build.