Updated the (incorrect) users of atlas_add_root_dictionary(...).
In the past there was a (very fragile) feature in atlas_add_root_dictionary(...)
that allowed it to scan the include definitions in a LinkDef.h file, and construct
the rootcint command based on that.

This feature has now been removed, as it was one of the most fragile parts
of our code. Now all users of atlas_add_root_dictionary(...) must declare explicitly
in CMake which headers they want to generate dictionaries for.