LArROD: Remove HV correction option from LArRawChannelBuilder.
An earlier change altered LArRawChannelBuilder to get the HV scale correction
from conditions rather than from the tool.  These are however not equivalent,
as the conditions version is already normalized to the online correction.

However, the scale correction of LArRawChannelBuilder is not enabled anywhere.
So rather than trying to sort this out now, simply remove the option.