Updating DigitizationAlg and PileUpToolsAlg to use private ToolHandles
This commit updates `DigitizationAlg` and `PileUpToolsAlg` to use
a private `ToolHandleArray<IPileUpTool>`. This will change the name
of the PileUpTools from `ToolSvc.PileUpToolGivenName` to
`AlgorithmName.PileUpToolGivenName`. This will change the digitization
output as the PileUpTool names are used to seed the random number
streams. Having the PileUpTool output depend on the parent Algorithm
name in this way is not ideal, so it may be necessary to find a
workaround in the future.

Because the PileUpTool values changed then the `test_MC16a_Digi_tf_configuration`
unit test had to be updated accordingly.