Implementing ms-only muon triggers
This mostly just sets up some pieces needed to run ms-only muon chains, including some refactoring of the sequences.
The ms-only chain is currently commented out because it doesn't run when some of the other chains are enabled. Having it implemented should allow easier debugging to get it running properly.