Fixed the installation of flake8_atlas.
The immediate problem was that with a recent AtlasCMake update
atlas_build_run.sh is now in a different place. But while fixing that,
I also had to realise that in the current setup flake8_atlas was not
installed alongside the release. It was only ever put into the
build directory. So I fixed that as well.