Preparing for updated small-R JER uncertainties
This MR adds some small functionality updates aimed at
small-R JER uncertainties.  In particular, the following is changed:

1. CalibArea is changed to CalibArea-07

2. MC and (pseudo-)data smearing seeds are now different

3. JER uncertainties now support MCTYPE generalization

In practice, this means:

1. The new unertainties will be picked up by default
(although users can always change the tool config to pick up
any "release" of uncertainties)

2. Users can now smear MC twice, once as MC and once as pseudo-data,
as the seeds are now distinct (no longer the same for both uses)

3. AFII and fullsim JER uncertainties are supported together,
instead of the past where only fullsim was supported

The corresponding inputs to CalibArea should come very soon, as I
have validated these changes with the potential inputs.  We are
just sorting out final validation of the inputs now.  So the
code changes for the upcoming recommendation are complete and in
this MR, while the input validation is ongoing.