TauTriggerMonitoring fix to handle AODSLIM content
Hello,

This MR implements a simple fix to solve the crash reported in ATLPHYSVAL-609 and ATR-19712.
Our monitoring was assuming that it would always run on AODFULL, and did not handle AODSLIM content.
We now check whether the HLT tau track and HLT cluster collections are present, this is needed for the monitoring of RNN ID inputs.

Cheers,
Bertrand