IOVSvc: Try to fix potential deadlock between StoreGate and IOVSvc.
IOVSvc::createCondObj locks IOVSvc and calls SGImplSvc::proxy, which
locks the store.

SGImplSvc::regFcn and friends lock the store and then call IOVSvc,
which tries to acquire its own lock.

Thus we have a potential deadlock.

Change IOVSvc::createCondObj so that it's not holding the IOVSvc lock
when it calls proxy().