Adding ID track thinning to tau track thinning tool
Hello,

This MR adds the possibility to keep the ID track particles that are associated with tau tracks.
Else, the TauTrack->track() link can end up broken.
In addition, the identification of relevant tracks is now based on the element link index, instead of the dR matching.
This new functionality will be needed by an upcoming MR of DFSUSY and DFTau.

Cheers,
Bertrand