CaloCalibHitRec: Preparing to make CaloDD immutable.
Previously, we were creating the CaloDD via a converter, then applying
alignments via conditions callbacks.  This this meant that the CaloDD
will be different depending on whether we're looking before or after
alignments have been applied.  This can then lead to problems if the
CaloDD is retrieved too early.  The const_cast'ing needed is also
problematic for MT.

Instead of building CaloDD via a converter, we want to move to creating
it explicitly as part of doing alignments.

However, this means that in general, we shouldn't be trying to retrieve
CaloDetDescrManager during initialize().  Rework to avoid this.