Apply HLT tau thinning only when trigger collections are present
Hello,

This is a fix to solve the ART test failure for SUSY3 and TAUP3.
The problem comes from the HLT tau thinning (!22992) which is meant to work only for post-TS1 data18 and special MC16e trigger samples.
While in ART, the data file is pre-TS1 and the MC file is AODSLIM, so there are some broken element links, and the thinning fails.
The implemented workaround consists in retrieving the run number for data, and the AMITag for MC, and enable the thinning only when relevant.
When I have time, I will improve the robustness of DerivationFrameworkSUSY/src/TauTracksThinning.cxx, and check the validity of element links before accessing them.

Cheers,
Bertrand