Remove MuidCaloEnergyMeas from muon trigger configuration
We never actually use the MuidCaloEnergyMeas tool in the trigger, since we don't have access to calo information at the muon trigger level. Any algorithms/tools that need it we anyways configure such that this tool is disabled.
Removing this from the trigger configuration so some cleanup of the offline code can proceed in https://gitlab.cern.ch/atlas/athena/merge_requests/23065.