try to address Athena CI failure
For some reason opening the histogram file failed for one of the three
jobs, complaining that the file is already there.  The best
explanation I have is some race condition between the three in opening
the file.  I now gave them all unique names in the hopes of resolving
that.