RHadrons Update for stops and sbottoms
This merge request fixes two sets of problems.

First, there were problems in the logic that assigned hadronic
interactions to R-hadrons. The problem was essentially that one quark in
the light quark system was ignored when calculating the R-hadron
properties. This is (surprisingly) a quite good approximation as long as
you aren't making extremely heavy R-hadrons, and as long as you don't
have stop or sbottom R-mesons, where there is only one light quark in
the R-hadron. That's fixed, the list of R-hadron interactions has
changed somewhat, and the results change a little bit.

Second, we needed a few minor adjustments to get stop and sbottom
R-hadron simulation working. That should all be done. Validation plots
have been made and everything.

This MR gets 21.6 back in sync with 21.0 following !23144