Fix thread-unsafety in EscapedEnergyRegistry class [ATLASSIM-4106]
EscapedEnergyRegistry class was implemented as a singleton, so it was not thread-safe.
This affected the HITs of LArCalibrationHitDeadMaterial when running in MT with the CaloCalibration on.
Added the #ifdef G4MULTITHREADED directive to handle
the multithreaded case. One instance of the class will be created
per each thread and stored in a tbb::concurrent_unordered_map that
is hashed with the threadID number.