SGComps: Have AddressRemappingSvc implement deletions via renaming.
When a clid/key is mentioned in a WriteHandleKey, we hide the corresponding
object from an input file.  This has been handled in AddressRemappingSvc by
removing any such TAD from the list of TADS.

However, it turns out to be useful in some cases to be able to get a proxy
for the original objects in the input file.  This is the case, for example,
when we want to wrap converters with algorithms.  Change so that instead
of deleting a TAD, we rename it by appending `_DELETED' to the key.
We do not preserve symlinks/aliases at this time.