Fix input track collections for building PRD association for TRT only tracks.
Assign the current state of the list of track collections rather than a
reference to the list of all track collections.

This removes track collections from the list of input track collections, which
are not supposed to be used for building the PRD association map for TRT only
tracks, and which do not exist at this stage.