Fix thinning of HLTResult (ATR-18627)
The thinning was never applied to the old-style `HLT::HLTResult` as it
is stored as `const` in the event store. Fix this by const-cast. The
thinning of the `xAOD::TrigNavigation` worked as it's already non-const
in the store.

Besides fixing the bug, removed one templated method as we need
different treatment for the const/non-const case.