AthenaCommon: Workaround issue mixing old/new configurations.
ToolSvc silently discards attempts to add duplicates.
But if configurableRun3Behavior is enabled, then those duplicates
may be distinct Configurable instances.  If we're also reading an old
configuration fragment, this can lead to errors about public tools
not being in ToolSvc.   Work around by forcing the owner of rejected
duplicates to be ToolSvc.  (Only enabled when configurableRun3Behavior is on.)