Revise Powheg SVN revision checking in PowhegControl
Now doesn't raise an exception if the .svn directory has been stripped from the Powheg process
directory installed on CVMFS. Instead, it will fall back to trying to read the revision from a
text file that lists them. If this fails as well, the revision is reported as 'unknown' without
error.