Prep for new RNN and DL1 in some jet collections
This lets us run the new versions of the RNN and DL1 instead of the old ones.

This code is nearly ready to be activated, but in the current state it should
not do anything at all. Once we uncomment some of the items in the
JetCollectionToTrainingMaps the AODFix will activate, and several btagging
algorithms will turn off.