Fix for ATLASRECTS-4909
Occasionally (once very 10k events or so) a low-quality, low-pT standalone muon cannot be succesfully extrapolated back through the calorimeter by the ParticleCaloExtensionTool.
Since this process is part of the isolation calculation, this leads to a crash in the IsolationBuilder.
In order to avoid these crashes, this fix attempts to obtain an extension for SA muons in the MuonCreatorTool.
If the extension cannot be calculated, or is empty, the muon is rejected.