TileConditions: Clean up TileInfoLader.
TileInfoLoader was playing games with the memory management of the TileInfo
object, removing it from the detector store at the end of a run and
adding it again at the start of a run.  None of this is needed, so clean
this up.  Also, TileInfo need not derive from DataObject and should get the
clid assigned properly with CLASS_DEF.