First step towards decoupling LArHV from the access to DCS conditions data in reco jobs
Modified EMBHVElectrode::hvLineNo() such that its reco clients now need to pass to it a pointer
to a LArHVIdMapping condition object, which provides mapping from electrodes to HV lines.

Migrated two clients of EMBHVElectrode - LArHVCondAlg and LArHVPathologyDbCondAlg - to the
new interface.

The AthSimulaton project was protected from the addition of new dependency on LArRecConditions
package, which hosts the LArHVIdMapping class, by leveraging the SIMULATIONBASE preprocessor statement