Increase character variable array sizes to avoid pessamistic compiler warnings
The compile was warning that the some arrays were smaller than the possible maximum
sizes of the strings that might be written to them. This was not a problem, since
the maximum allocation was already larger than what could reasonably be used, but the
sizes have been increased accordingly so this warning should no longer occur.