Fixes to RHadron simulation
Most fixes affect the simulation of squark R-hadrons, but some affect
gluino R-hadrons as well. The function to calculate baryon number was
just wrong (only gave +1 or 0, and didn't correctly handle several
classes of particles). The calculation of strangeness was also
inconsistent.

All fixed now, validated, and tested through to plots of the simulation
that have been made public in ATL-PHYS-PUB-2019-019