Merge request to sync master with 21.3
The latest changes to 21.3 that should have been swept automatically failed to sweep because of an out of date file in master. So these changes just include the changes for the files from !23630 (merged)

In addition, there are a lot of other changes that appeared to already be there due to !23613 (merged) in the from !23630 (merged), but in the fresh checkout of master that I made just a minute ago, they still seem to be missing. As a consquence could this just be accepted -all these changes have already been reviewed and this request is simply to force the files to be updated since the automatic sweeps failed.