TrigServices: Enable static thread checker
Mark a few methods as non-thread safe. `TrigCOOLUpdateHelper` cannot use
the recommended way to access conditions via a `ReadCondHandle` as the
information needs to be read before the first event. However, the COOL
folder (/TRIGGER/HLT/COOLUPDATE) never changes during the run and thus
it is safe to use the old-style `DataHandle` access to the conditions
object.

Also apply some general cleanup to the code using C++17 features.