Fix nJet20 moment calculation
Now, nJet20 will always be the same regardless of the calibration
applied to the trigger jet.

This breaks backward compatibility but this was not part of any
recommendation. This is just being used to derive calibrations for
trigger jets for the TLA anaysis. This has to be fixed to derive new
calibrations.